+++ __tests__/html/typing/informative.js
@@ -1,5 +1,0 @@
-/** @jest-environment ./packages/test/harness/src/host/jest/WebDriverEnvironment.js */
-
-describe('informative typing message', () => {
-  test('should be shown as typing indicator', () => runHTML('typing/informative'));
-});